pyproject.toml
[tool.poetry]
name = "geopy"
version = "0.1.0"
description = ""
authors = ["Mike Romanovsky"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.3"
fastapi = "^0.113.0"
sqlalchemy = "^2.0.34"
pydantic = {extras = ["email"], version = "^2.9.1"}
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
asyncpg = "^0.29.0"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
bcrypt = "^4.2.1"
pydantic-extra-types = "^2.10.1"
websockets = "^15.0"
python-multipart = "^0.0.20"
email-validator = "^2.2.0"
redis = "^5.2.1"
py-spy = "^0.4.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"
httpx = "^0.27.2"
mypy = "^1.11.2"
pytest-timeout = "^2.3.1"
memory-profiler = "^0.61.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true

[[tool.mypy.overrides]]
module = "migrations/env.py"
ignore_missing_imports = true
